AJAY WAGH - QUIZ GENERATING AGENT/requirements.txt
Flask
flask-cors
langchain-google-genai
langchain-core
langchain
python-dotenv
gunicorn
Pillow